+++ client/src/components/Header/index.tsx
@@ -5,7 +5,7 @@ import { useRouter } from "next/router"
 import Logo from '../../assets/img/logo.svg'
 import { useAppSelector } from "../../hooks/useAppSelector"
 import CartButton from "../CartButton"
-import Search from "../Search"
+import Search from "./Search"
 
 const Header = () => {
