+++ client/src/components/screens/CartPage/index.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const CartPage = () => {
+    return (
+        <div>
+            
+        </div>
+    );
+};
+
+export default CartPage;
